\name{RecAsso}
\alias{RecAsso}
\docType{package}
\title{Recursive mining for frequent itemSet and exporting Rules.}
\description{This  version makes it possible to carry out the usual applications of association rule Mining, and propose some other possibilities to define Rules depending on the usecase .This by using recursive algorithms. All of the functions are using the C++ language.  It will be updated soon with functions allowing it to be applied on non-fixed databases.}


